fixed runtime error with import.meta in WebWorker
The loudness WebWorker imported the loudness module, which also includes a function for running a WebWorker and therefore includes a call to `import.meta.url`, which is not allowed in WebWorker.
The solution is to extract the necessary code of the loudness module into a new and separated loudness-calculations module that is then imported by the WebWorker and the loudness module.